Add Badge to answer channel, refactoring Answer channel, fixed Files, Links
