+++ app/javascript/utilities/templates/question/create_question.js
@@ -29,5 +29,6 @@ function questionCreate(data) {
       questionDiv.append(buttonDelete);
     }      
     questionList.prepend(questionDiv);
+    return questionList
   }
 }
