--- app/views/Convert/shared/_errors.html.slim
@@ -1,5 +1,0 @@
-- if resource&.errors.present?
-  b= "#{resource.errors.count} error(s) detected: "
-  ul
-    -resource.errors.full_messages.each do |message|
-      li= message
