--- app/views/Convert/shared/_link.html.slim
@@ -1,18 +1,0 @@
-- resource.links.each do |link|
-  li id="link_#{link.id}"
-    - if link.gist?
-      .pb-2.pt-3.pr-1.d-flex.justify-content-between
-        .link-content
-          = link.name
-          button.btn.btn-outline-light.mx-2.gist-button style=("background: darkgray;")  Show gist
-          #gist_link.hide data-gistlink-id="#{link.id}" data-src-id="#{parse_gist_id(link.url)}"
-            #gistFrame 
-              = render html: link.url.html_safe
-        .link-delete
-          - if user_signed_in? && link.persisted?
-            = link_to "Delete Link", link_path(link), method: :delete, remote: true
-    - else
-      .pb-2.pt-3.pr-1.d-flex.justify-content-between 
-        = link_to link.name, link.url
-        - if user_signed_in? && link.persisted?
-          = link_to "Delete Link", link_path(link), method: :delete, remote: true
