--- app/views/Convert/shared/_link_fields.html.slim
@@ -1,9 +1,0 @@
-/*- binding.pry
-.nested-fields id="link-form-#{f.object.id if f.object.present?}"
-  = f.hidden_field :_destroy
-  Name  
-    = f.input :name, label: false, required: false, input_html: { class: "form-control" }, placeholder: 'Name'
-  URL
-    = f.input :url, label: false, required: false, input_html: { class: "form-control mt-2" }, placeholder: 'URL'
-  
-  = link_to "Remove link", '#', class: "remove_fields"
