--- app/views/Convert/shared/_navigation.html.slim
@@ -1,14 +1,0 @@
-.container-fluid
-  nav.navbar.navbar-expand-lg.navbar-light.bg-light.fw-lighter
-    .navbar-nav.ml-auto.px-3
-      = link_to 'Home', root_path, data: {item: 'Home'}, class: "#{'current' if current_page?(root_path)}"
-    .navbar-nav.mr-auto.px-3
-      - if user_signed_in?
-        = link_to current_user.email, edit_user_registration_path, id: 'account', class: "#{ 'admin' if current_user.admin? }", data: {item: current_user.email}
-        - if current_user.rewards.present?
-          = link_to rewards_path, class: "nav-item nav-link rewards-icon" do
-            i.fas.fa-award
-        = link_to "Logout", destroy_user_session_path, method: :delete, data: { item: 'Logout' }
-      - else
-        = link_to 'Login', new_user_session_path, data: { item: 'Login' }
-        = link_to "Logon", new_user_registration_path, data: { item: 'Logon' }, class: 'logon'
