--- app/views/Convert/shared/_vote.html.slim
@@ -1,41 +1,0 @@
-div class="vote_content-answer_#{@naswer.id}"
-  div class="vote-answer_#{@answer.id} d-flex justify-content-between"
-    .vote-info.d-flex.justify-content-end.mt-3.mb-2
-      span class="votes-answer_#{@answer.id}-total votes-total" 
-        = "Rating: @answer.rating"
-    - if policy(@answer).make_vote?
-      .vote-action.d-flex.justify-content-start.mt-3.mb-2
-        - if policy(@answer).upvote?
-          = link_to polymorphic_path(@answer, action: :upvote), 
-                    class: "upvote upvote-#{ident(@answer)}_#{@answer.id}", 
-                    method: :post,
-                    data: { type: :json, setobject: "#{@answer}" },
-                    remote: true, 
-                    title: 'Vote ▲' do
-            span.text-success.fs-5.fa.fa-thumbs-up
-        - else
-          span.text-muted.op-40.fs-5.fa.fa-thumbs-up
-
-        - if policy(@answer).resetvote?
-          = link_to polymorphic_path(@answer, action: :resetvote),
-                  class: "reset reset-#{ident(@answer)}_#{@answer.id}",
-                  method: :delete,
-                  data: { type: :json },
-                  remote: true,
-                  title: 'Vote reset' do
-            span.text-secondary.fs-5.fa.fa-times
-        - else
-          span.text-muted.op-40.fs-5.fa.fa-times
-
-        - if policy(@answer).downvote?
-          = link_to polymorphic_path(@answer, action: :downvote),
-                    class: "downvote downvote-#{ident(@answer)}_#{@answer.id}",
-                    method: :post,
-                    data: { type: :json },
-                    remote: true,
-                    title: 'Vote ▼' do
-            span.text-danger.fs-5.fa.fa-thumbs-down
-        - else
-          span.text-muted.op-40.fs-5.fa.fa-thumbs-down
-  div class="vote-errors-#{ident(@answer)}_#{@answer.id} justify-content-center text-danger py-1 d-none"
-
