--- app/views/Convert/shared/create_comment.js.erb
@@ -1,24 +1,0 @@
-var errorsComment = {}
-errorsComment = document.querySelector('.comments-error-<%= ident(@commentable) %>_<%= @commentable.id %>')
-
-// Timeout hide inline error form 
-function fadeOut(object) {
-  if (object) {
-    object.classList.add('fadeout');
-    window.setTimeout(function() {
-      object.innerText = '';
-    }, 7000);
-  }
-}
-
-if (errorsComment) { errorsComment.innerText = '' }
-
-<% if @comment.errors.present? %>  
-  errorsComment.insertAdjacentHTML('beforeend', '<%= errors_for(@comment) %>');
-  fadeOut(errorsComment)
-<% else %>
-  <% return unless @comment.persisted? %>
-  document.getElementById('<%= ident(@commentable) %>-comments-<%= @commentable.id %>').innerText = '';
-  document.getElementById('<%= ident(@commentable) %>-comments-<%= @commentable.id %>').insertAdjacentHTML('beforeend', "<%= j render @commentable.comments.includes(:user) %>"); 
-  document.getElementById('comments-<%= ident(@commentable) %>_<%= @commentable.id %>').reset();
-<% end %>
