--- app/views/Convert/shared/delete_comment.js.erb
@@ -1,4 +1,0 @@
-let comment_<%= @comment.id %> = document.querySelector('#comment_<%= @comment.id %>');
-if (comment_<%= @comment.id %>) {
-  comment_<%= @comment.id %>.remove();
-}
