--- app/views/Convert/shared/output.html.erb
@@ -1,1 +1,0 @@
-
