--- app/views/Convert/shared/output1.html.erb
@@ -1,41 +1,0 @@
-<div class="vote_content-answer_<%= ::Temple::Utils.escape_html((@naswer.id)) %>">
-<div class="vote-answer_<%= ::Temple::Utils.escape_html((@answer.id)) %> d-flex justify-content-between">
-<div class="vote-info d-flex justify-content-end mt-3 mb-2">
-<span class="votes-answer_<%= ::Temple::Utils.escape_html((@answer.id)) %>-total votes-total">
-<%= ::Temple::Utils.escape_html(("Rating: @answer.rating")) %>
-</span></div><% if policy(@answer).make_vote?
- %><div class="vote-action d-flex justify-content-start mt-3 mb-2">
-<% if policy(@answer).upvote?
-_slim_controls1 = link_to polymorphic_path(@answer, action: :upvote),
-class: "upvote upvote-#{ident(@answer)}_#{@answer.id}",
-method: :post,
-data: { type: :json, setobject: "#{@answer}" },
-remote: true,
-title: 'Vote ▲' do %><% _slim_controls2 = ''; 
-; _slim_controls2 << ("<span".freeze); _slim_controls2 << (" class=\"".freeze); _slim_controls2 << ("text-success".freeze); _slim_controls2 << (" ".freeze); _slim_controls2 << ("fs-5".freeze); _slim_controls2 << (" ".freeze); _slim_controls2 << ("fa".freeze); _slim_controls2 << (" ".freeze); _slim_controls2 << ("fa-thumbs-up".freeze); _slim_controls2 << ("\"".freeze); _slim_controls2 << (">".freeze); 
-; _slim_controls2 << ("</span>".freeze); _slim_controls2 %><% end %><%= ::Temple::Utils.escape_html((_slim_controls1)) %><% else
- %><span class="text-muted op-40 fs-5 fa fa-thumbs-up">
-
-</span><% end; if policy(@answer).resetvote?
-_slim_controls3 = link_to polymorphic_path(@answer, action: :resetvote),
-class: "reset reset-#{ident(@answer)}_#{@answer.id}",
-method: :delete,
-data: { type: :json },
-remote: true,
-title: 'Vote reset' do %><% _slim_controls4 = ''; 
-; _slim_controls4 << ("<span".freeze); _slim_controls4 << (" class=\"".freeze); _slim_controls4 << ("text-secondary".freeze); _slim_controls4 << (" ".freeze); _slim_controls4 << ("fs-5".freeze); _slim_controls4 << (" ".freeze); _slim_controls4 << ("fa".freeze); _slim_controls4 << (" ".freeze); _slim_controls4 << ("fa-times".freeze); _slim_controls4 << ("\"".freeze); _slim_controls4 << (">".freeze); 
-; _slim_controls4 << ("</span>".freeze); _slim_controls4 %><% end %><%= ::Temple::Utils.escape_html((_slim_controls3)) %><% else
- %><span class="text-muted op-40 fs-5 fa fa-times">
-
-</span><% end; if policy(@answer).downvote?
-_slim_controls5 = link_to polymorphic_path(@answer, action: :downvote),
-class: "downvote downvote-#{ident(@answer)}_#{@answer.id}",
-method: :post,
-data: { type: :json },
-remote: true,
-title: 'Vote ▼' do %><% _slim_controls6 = ''; 
-; _slim_controls6 << ("<span".freeze); _slim_controls6 << (" class=\"".freeze); _slim_controls6 << ("text-danger".freeze); _slim_controls6 << (" ".freeze); _slim_controls6 << ("fs-5".freeze); _slim_controls6 << (" ".freeze); _slim_controls6 << ("fa".freeze); _slim_controls6 << (" ".freeze); _slim_controls6 << ("fa-thumbs-down".freeze); _slim_controls6 << ("\"".freeze); _slim_controls6 << (">".freeze); 
-; _slim_controls6 << ("</span>".freeze); _slim_controls6 %><% end %><%= ::Temple::Utils.escape_html((_slim_controls5)) %><% else
- %><span class="text-muted op-40 fs-5 fa fa-thumbs-down">
-</span><% end %></div><% end %></div><div class="vote-errors-<%= ::Temple::Utils.escape_html((ident(@answer))) %>_<%= ::Temple::Utils.escape_html((@answer.id)) %> justify-content-center text-danger py-1 d-none">
-</div></div>
