--- app/views/Convert/shared/slim2erb.sh
@@ -1,5 +1,0 @@
-for x in $(find . -name \*.slim); do
-  slimrb --rails -e $x $(echo "$x" | sed 's/\.slim$/.erb/')
-  echo $x
-done
-find . -name "*.slim" -type f -delete
