+++ app/views/questions/_question_for_bc.html.slim
